Refactor schema management: update import paths, remove unused schemas, and implement new schema creation components
- Updated import path for schema utilities in base.worker.ts
- Renamed renderNodePage to renderPage in DefaultScenarioNode
- Deleted unused schemas.tsx file and moved relevant code to schemas.tsx
- Added createSchemaComponents.tsx for schema creation UI components
- Implemented schemaService.ts for handling schema-related API calls
- Introduced validation utilities for schema forms and grid layers
- Enhanced SchemasPage component to manage schema creation and validation
- Added types for schema components and validation results
