+++ src/framework/src/components/functionPage/types.ts
@@ -1,7 +1,5 @@
-import { ISceneTree } from "@/core/scene/iscene"
+import { ISceneNode } from '@/core/scene/iscene'
 
 export type CreatePageProps = {
-    creationType: 'schema' | 'patch',
-    resourceTree: ISceneTree,
-    onCreationSuccess: (resourceTree: ISceneTree, creationType: 'schema' | 'patch') => void
+    node: ISceneNode
 }
