+++ src/framework/src/resource/scenarioRegistry.tsx
@@ -1,7 +1,7 @@
 import DefaultScenarioNode, { DefaultPageContext } from './scenario/default'
 import RootScenarioNode from './scenario/root'
 import TopoScenarioNode from './scenario/topo'
-import SchemasScenarioNode, { SchemasPageContext } from './scenario/schemas'
+import SchemasScenarioNode, { SchemasPageContext } from './scenario/schemas/schemas'
 import SchemaScenarioNode from './scenario/schema'
 import PatchesScenarioNode, { PatchesPageContext } from './scenario/patches'
